+++ packages/eslint-config/testing/jest-dom.js
@@ -1,0 +1,7 @@
+const merge = require('lodash.merge');
+
+const jest = require('./jest');
+
+module.exports = merge(jest, {
+  extends: ['plugin:jest-dom/recommended'],
+});